Restore cgroup_memory_recap, and move new functionality into cgroup_perf_recap
